Move the twisted utility methods to its own module.
To avoid problems when not running the reactor on a separate thread.
