+++ Tribler/Core/Utilities/twisted_thread.py
@@ -179,20 +179,3 @@ def deferred(timeout=None):
         wrapper = make_decorator(func)(wrapper)
         return wrapper
     return decorate
-
-
-def callInThreadPool(fun, *args, **kwargs):
-    """
-    Calls fun(*args, **kwargs) in the reactor's thread pool.
-    """
-    if isInThreadPool():
-        fun(*args, **kwargs)
-    else:
-        reactor.callFromThread(reactor.callInThread, fun, *args, **kwargs)
-
-
-def isInThreadPool():
-    """
-    Check if we are currently on one of twisted threadpool threads.
-    """
-    return current_thread() in reactor.threadpool.threads
